Moved hamilton tools to compat so they can be used more widely
