+++ compat/hamilton-tools.cpp
