+++ compat/hamilton-tools.h
